adjusted the frontend for changes in the api
